+++ front/src/views/RegularPage/RegularPage.tsx
@@ -3,10 +3,10 @@ import ReactMarkdown from 'react-markdown';
 
 import remarkGfm from 'remark-gfm';
 
-import { Container } from '../../components/Container';
 import { Gainburua } from '../../components/Gainburua';
 import { Oina } from '../../components/Oina';
 import { RegularPageContent } from '../../domain/basicPage/RegularPageContent';
+import { Container } from '../components/Container';
 import * as styles from './RegularPage.module.scss';
 
 interface Props {
