+++ front/src/views/components/Container/Container.module.scss
@@ -1,4 +1,4 @@
-@import '../../styles/tools/mediaQueries.scss';
+@import '../../../styles/tools/mediaQueries.scss';
 
 .wrapper {
   margin-left: auto;
